Make autoloading more standard using classmaps
