+++ test/CommonMarkTestWeak.php
@@ -1,5 +1,4 @@
 <?php
-require_once(__DIR__ . '/CommonMarkTestStrict.php');
 
 /**
  * Test Parsedown against the CommonMark spec, but less aggressive
