+++ test/ParsedownTest.php
@@ -1,5 +1,4 @@
 <?php
-require 'SampleExtensions.php';
 
 use PHPUnit\Framework\TestCase;
